# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Urva Gandhi Portfolio Developer API & MCP Server",
    "description": "Public REST APIs, Model Context Protocol (MCP) server, and developer endpoints providing verified coding statistics, developer profile metadata, and interactive contact handling for Urva Gandhi.",
    "version": "1.0.0",
    "contact": {
      "name": "Urva Gandhi",
      "email": "23bce078@nirmauni.ac.in",
      "url": "https://urvagandhi.tech"
    },
    "x-versioning-policy": "URL path versioning (/v1/) and X-API-Version header parameter. Backward compatibility guaranteed for version 1.",
    "x-deprecation-policy": "Deprecation notices served via Sunset HTTP header 6 months prior to end-of-life."
  },
  "servers": [
    {
      "url": "https://urvagandhi.tech",
      "description": "Primary Production Domain"
    },
    {
      "url": "https://urvagandhi-portfolio.vercel.app",
      "description": "Lifetime Vercel Mirror Domain"
    },
    {
      "url": "https://urvagandhi.tech/v1",
      "description": "Active Domain Versioned Base URL (/v1)"
    }
  ],
  "paths": {
...
    "/api/mcp": {
      "get": {
        "operationId": "getMcpManifest",
        "summary": "MCP Server Handshake & Capabilities",
        "description": "Returns Model Context Protocol server capabilities, tool declarations, and server metadata.",
        "security": [],
        "responses": {
          "200": {
            "description": "MCP Manifest payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpManifest"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          }
        ]
      },
      "post": {
        "operationId": "callMcpTool",
        "summary": "Execute MCP Tool or Handshake (JSON-RPC 2.0)",
        "description": "Handles MCP JSON-RPC 2.0 requests (initialize, tools/list, tools/call, resources/list, resources/read, ping).",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          }
        ]
      }
    },
...
    "/v1/mcp": {
      "get": {
        "operationId": "getMcpManifest",
        "summary": "MCP Server Handshake & Capabilities",
        "description": "Returns Model Context Protocol server capabilities, tool declarations, and server metadata.",
        "security": [],
        "responses": {
          "200": {
            "description": "MCP Manifest payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpManifest"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          }
        ]
      },
      "post": {
        "operationId": "callMcpTool",
        "summary": "Execute MCP Tool or Handshake (JSON-RPC 2.0)",
        "description": "Handles MCP JSON-RPC 2.0 requests (initialize, tools/list, tools/call, resources/list, resources/read, ping).",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          }
        ]
      }
    },
    "/api/v1/mcp": {
      "get": {
        "operationId": "getMcpManifest",
        "summary": "MCP Server Handshake & Capabilities",
        "description": "Returns Model Context Protocol server capabilities, tool declarations, and server metadata.",
        "security": [],
        "responses": {
          "200": {
            "description": "MCP Manifest payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpManifest"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          }
        ]
      },
      "post": {
        "operationId": "callMcpTool",
        "summary": "Execute MCP Tool or Handshake (JSON-RPC 2.0)",
        "description": "Handles MCP JSON-RPC 2.0 requests (initialize, tools/list, tools/call, resources/list, resources/read, ping).",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://urvagandhi.tech/.well-known/oauth/token",
            "scopes": {
              "profile:read": "Read public developer profile and coding statistics.",
              "profile:write": "Submit a contact inquiry on behalf of an authenticated client."
            }
          },
          "authorizationCode": {
            "authorizationUrl": "https://urvagandhi.tech/.well-known/oauth/authorize",
            "tokenUrl": "https://urvagandhi.tech/.well-known/oauth/token",
            "scopes": {
              "profile:read": "Read public developer profile and coding statistics.",
              "profile:write": "Submit a contact inquiry on behalf of an authenticated client."
            }
          }
        }
      }
    },
    "schemas": {
...
      "McpManifest": {
        "type": "object",
        "properties": {
          "schema_version": {
            "type": "string",
            "example": "1.0"
          },
          "name": {
            "type": "string",
            "example": "urva-gandhi-portfolio-mcp"
          }
        }
      },
      "McpRequest": {
        "type": "object",
        "required": ["jsonrpc", "method"],
        "properties": {
          "jsonrpc": {
            "type": "string",
            "example": "2.0"
          },
          "id": {
            "type": "integer",
            "example": 1
          },
          "method": {
            "type": "string",
            "example": "tools/call"
          },
          "params": {
            "type": "object"
          }
        }
      },
      "McpResponse": {
        "type": "object",
        "properties": {
          "jsonrpc": {
            "type": "string",
            "example": "2.0"
          },
          "id": {
            "type": "integer",
            "example": 1
          },
          "result": {
            "type": "object"
          }
        }
      }
    }
  }
}