# src/pages/api/mcp.js
import mcpManifest from "../../../public/.well-known/mcp.json";

const PRIMARY_DOMAIN = "https://urvagandhi.tech";
const LIFETIME_DOMAIN = "https://urvagandhi-portfolio.vercel.app";

const SUPPORTED_PROTOCOL_VERSIONS = ["2024-11-05", "2025-03-26", "2025-06-18"];
const DEFAULT_PROTOCOL_VERSION = "2024-11-05";

const jsonrpcResponse = (id, resultOrError) => {
  if (
    resultOrError &&
    Object.prototype.hasOwnProperty.call(resultOrError, "error")
  ) {
    return { jsonrpc: "2.0", id, ...resultOrError };
  }
  return { jsonrpc: "2.0", id, result: resultOrError };
};

export default async function handler(req, res) {
  const host = req.headers.host || "urvagandhi.tech";
  const proto =
    req.headers["x-forwarded-proto"] ||
    (host.includes("localhost") ? "http" : "https");
  const activeDomain = `${proto}://${host}`;

  // Standard RFC Rate Limit Headers & Agentic Discovery Headers
  res.setHeader("Access-Control-Allow-Origin", "*");
  res.setHeader("Access-Control-Allow-Methods", "GET, POST, OPTIONS");
  res.setHeader(
    "Access-Control-Allow-Headers",
    "Content-Type, Authorization, Accept, Mcp-Version, Mcp-Protocol-Version, X-API-Version",
  );
  res.setHeader("Mcp-Version", "1.0");
  res.setHeader("Mcp-Protocol-Version", "2024-11-05");
  res.setHeader("X-API-Version", "1.0.0");
  res.setHeader("RateLimit-Limit", "100");
  res.setHeader("RateLimit-Remaining", "99");
  res.setHeader("RateLimit-Reset", "60");
  res.setHeader("X-RateLimit-Limit", "100");
  res.setHeader("X-RateLimit-Remaining", "99");
  res.setHeader("X-RateLimit-Reset", "60");
  res.setHeader("Link", `<${activeDomain}/api/mcp>; rel="mcp-server"`);
  res.setHeader("Vary", "Accept, Accept-Encoding, Host");

  const dynamicManifest = {
    ...mcpManifest,
    homepage: activeDomain,
    primary_homepage: PRIMARY_DOMAIN,
    alternate_homepage: LIFETIME_DOMAIN,
    endpoints: {
      mcp: `${activeDomain}/api/mcp`,
      sse: `${activeDomain}/api/mcp`,
      primary_mcp: `${PRIMARY_DOMAIN}/api/mcp`,
      lifetime_mcp: `${LIFETIME_DOMAIN}/api/mcp`,
    },
    servers: [
      {
        url: `${activeDomain}/api/mcp`,
        description: "Active Environment Server Endpoint",
      },
      {
        url: `${PRIMARY_DOMAIN}/api/mcp`,
        description: "Primary Custom Domain Endpoint",
      },
      {
        url: `${LIFETIME_DOMAIN}/api/mcp`,
        description: "Lifetime Vercel Mirror Endpoint",
      },
    ],
  };

  const wantsSse = (req.headers.accept || "")
    .split(",")
    .map((a) => a.trim().toLowerCase())
    .includes("text/event-stream");

  // Send a JSON-RPC 2.0 response honoring the Streamable HTTP transport
  // (SSE framing) when the client asks for text/event-stream.
  const sendJsonRpc = (id, resultOrError, status = 200) => {
    const payload = JSON.stringify(jsonrpcResponse(id, resultOrError));
    if (wantsSse) {
      res.setHeader("Content-Type", "text/event-stream");
      res.setHeader("Cache-Control", "no-cache, no-transform");
      res.setHeader("Connection", "close");
      res.status(status);
      return res.end(`data: ${payload}\n\n`);
    }
    res.setHeader("Content-Type", "application/json");
    return res.status(status).json(jsonrpcResponse(id, resultOrError));
  };

  if (req.method === "OPTIONS") {
    return res.status(200).end();
  }

  if (req.method === "GET") {
    return res.status(200).json(dynamicManifest);
  }

  if (req.method === "POST") {
    const body = req.body || {};
    const { jsonrpc, id, method, params } = body;

    const profileData = {
      name: "Urva Yogeshkumar Gandhi",
      title: "Software & AI Systems Engineer",
      domains: {
        active: activeDomain,
        primary: PRIMARY_DOMAIN,
        lifetime: LIFETIME_DOMAIN,
      },
      university: "Institute of Technology, Nirma University",
      degree: "B.Tech Computer Science & Engineering (Minor: Adaptive AI)",
      cgpa: "8.83 / 10",
      primarySkills: [
        "Java",
        "Spring Boot",
        "Spring Security",
        "REST APIs",
        "ReactJS",
        "Next.js",
        "JavaScript",
        "TypeScript",
        "Python",
        "C++",
        "PostgreSQL",
        "MongoDB",
        "MySQL",
        "Docker",
        "Git",
        "Data Structures & Algorithms",
        "Full Stack Development",
        "Backend Development",
        "Software Engineering",
        "Machine Learning",
        "Artificial Intelligence",
        "NLP",
        "Microservices",
        "OOP",
        "JWT",
        "OAuth 2.0",
      ],
      hackathons: [
        "1st Place – Health AI Innovation Hackathon 2025 (140+ teams)",
        "Track Runner-Up – HACKaMINeD 2026 (400+ teams)",
        "2nd Runner-Up – Odoo x Gujarat Vidyapith Hackathon 2026",
      ],
      location:
        "S-308, Venus Parkland, Near Vejalpur Police Chowki, Vejalpur, Ahmedabad, Gujarat, India - 380051",
      emails: ["23bce078@nirmauni.ac.in", "urvagandhi24@gmail.com"],
      phone: "+91-8866241204",
      github: "https://github.com/urvagandhi",
      linkedin: "https://www.linkedin.com/in/urva-gandhi/",
    };

    const llmsSummaryText = `# Urva Yogeshkumar Gandhi - Developer & Agent Context
Final year CSE undergraduate at Nirma University (CGPA 8.83/10, Minor in Adaptive AI).
Backend Engineer specializing in Java, Spring Boot, Microservices, REST APIs, and AI Agentic systems.
Primary Web Domain: ${PRIMARY_DOMAIN} | Alternate Mirror: ${LIFETIME_DOMAIN}`;

    // Handle MCP JSON-RPC 2.0 requests
    if (jsonrpc === "2.0" || method) {
      if (method && method.startsWith("notifications/")) {
        // JSON-RPC notifications are fire-and-forget: the server MUST NOT
        // return a response body. Acknowledge with 202 Accepted per the MCP
        // Streamable HTTP transport spec.
        res.status(202).end();
        return;
      }

      if (method === "initialize") {
        const requestedVersion = params?.protocolVersion;
        const negotiatedVersion = SUPPORTED_PROTOCOL_VERSIONS.includes(
          requestedVersion,
        )
          ? requestedVersion
          : DEFAULT_PROTOCOL_VERSION;
        res.setHeader("Mcp-Protocol-Version", negotiatedVersion);
        return sendJsonRpc(id, {
          protocolVersion: negotiatedVersion,
          capabilities: {
            tools: { listChanged: false },
            resources: { subscribe: false, listChanged: false },
          },
          serverInfo: {
            name: "urva-gandhi-portfolio-mcp",
            version: "1.0.0",
          },
          instructions:
            "Urva Gandhi portfolio MCP server. Public read-only tools: get_developer_profile, get_coding_stats, get_projects, get_contact_info. Available resources: portfolio://developer-profile and portfolio://llms-summary.",
        });
      }

      if (method === "ping") {
        return sendJsonRpc(id, {});
      }

      if (method === "tools/list") {
        return sendJsonRpc(id, {
          tools: mcpManifest.tools,
        });
      }

      if (method === "resources/list") {
        return sendJsonRpc(id, {
          resources: [
            {
              uri: "portfolio://developer-profile",
              name: "Urva Gandhi Developer Profile & Credentials",
              description:
                "Complete developer profile, Nirma University education, Java/Spring Boot & AI tech stack, and hackathon achievements.",
              mimeType: "application/json",
            },
            {
              uri: "portfolio://llms-summary",
              name: "LLM Agent Portfolio Summary",
              description:
                "Machine-readable summary of Urva Gandhi for AI agents.",
              mimeType: "text/markdown",
            },
          ],
        });
      }

      if (method === "resources/read") {
        const uri = params?.uri;
        let contentText = "";
        let mimeType = "text/plain";

        if (uri === "portfolio://developer-profile") {
          contentText = JSON.stringify(profileData, null, 2);
          mimeType = "application/json";
        } else if (uri === "portfolio://llms-summary") {
          contentText = llmsSummaryText;
          mimeType = "text/markdown";
        } else {
          return sendJsonRpc(
            id,
            {
              error: {
                code: -32602,
                message: `Resource not found: ${uri}`,
              },
            },
            200,
          );
        }

        return sendJsonRpc(id, {
          contents: [
            {
              uri: uri,
              mimeType: mimeType,
              text: contentText,
            },
          ],
        });
      }

      if (method === "tools/call") {
        const name = params?.name;
        const args = params?.arguments || {};

        let toolResult = null;

        if (name === "get_developer_profile") {
          toolResult = profileData;
        } else if (name === "get_coding_stats") {
          toolResult = {
            platform: args.platform || "all",
            platforms: {
              leetcode: {
                username: "Urva_Gandhi",
                solved: "300+ DSA Problems",
                rating: 1637,
                topPercentile: "Top 19.46%",
                profile: "https://leetcode.com/u/Urva_Gandhi/",
              },
              codeforces: {
                username: "Urva_Gandhi",
                profile: "https://codeforces.com/profile/Urva_Gandhi",
              },
              codechef: {
                username: "urva_gandhi",
                profile: "https://www.codechef.com/users/urva_gandhi",
              },
              gfg: {
                username: "urvagandhi",
                profile: "https://geeksforgeeks.org/user/urvagandhi",
              },
              hackerrank: {
                username: "urvagandhi",
                profile: "https://hackerrank.com/urvagandhi",
              },
            },
          };
        } else if (name === "get_projects") {
          toolResult = [
            {
              title: "CoinTrack – Multi-Broker Finance Platform",
              role: "Self Project (Team Size: 1)",
              techStack: [
                "Java",
                "Spring Boot",
                "Spring Security",
                "MongoDB",
                "Docker",
                "REST API",
                "JWT",
                "AES-256 Encryption",
                "Caffeine",
                "Bucket4j",
              ],
              description:
                "Designed & shipped a multi-broker investment platform with 45+ REST endpoints using hexagonal architecture. Aggregated Zerodha, Angel One, Upstox into real-time P&L dashboard. TOTP 2FA, OAuth 2.0, AES encryption.",
              github: "https://github.com/urvagandhi/cointrack",
            },
            {
              title: "FleetFlow – Fleet & Logistics Management System",
              role: "Odoo Hackathon 2026 (2nd Runner-Up)",
              techStack: [
                "ReactJS",
                "TypeScript",
                "ExpressJS",
                "PostgreSQL",
                "Prisma ORM",
                "Socket.IO",
                "Docker",
                "REST API",
              ],
              description:
                "Developed a full-stack logistics management system with 30+ APIs, RBAC, audit logging, and real-time updates via Socket.IO.",
              github: "https://github.com/urvagandhi/Odoo-Hackathon-26",
            },
            {
              title: "Agent Paperpal – Agentic AI Manuscript Formatter",
              role: "HACKaMINeD 2026 (Track Runner-Up)",
              techStack: [
                "Python",
                "CrewAI",
                "ReactJS",
                "Google Gemini",
                "Office.js",
                "AI Agents",
                "NLP",
              ],
              description:
                "Multi-agent AI platform automating manuscript formatting & compliance checking (APA, IEEE, Springer, Vancouver, Chicago).",
              github: "https://github.com/urvagandhi/TBD",
            },
            {
              title: "RWEsearch – Healthcare Analytics Platform",
              role: "Health AI Innovation Hackathon 2025 (1st Place Winner)",
              techStack: [
                "Python",
                "Scikit-learn",
                "XGBoost",
                "TensorFlow",
                "Machine Learning",
                "Data Analytics",
                "Streamlit",
              ],
              description:
                "Healthcare analytics platform predicting hospital readmissions using XGBoost & Deep Learning models with an interactive Streamlit UI.",
              github: "https://github.com/urvagandhi/RWEsearch-Hackathon",
            },
          ];
        } else if (name === "get_contact_info") {
          toolResult = {
            domains: {
              active: activeDomain,
              primary: PRIMARY_DOMAIN,
              lifetime: LIFETIME_DOMAIN,
            },
            emails: ["23bce078@nirmauni.ac.in", "urvagandhi24@gmail.com"],
            phoneNumbers: ["+91-8866241204", "+91-7203030498"],
            github: "https://github.com/urvagandhi",
            linkedin: "https://www.linkedin.com/in/urva-gandhi/",
            website: activeDomain,
            location:
              "S-308, Venus Parkland, Near Vejalpur Police Chowki, Vejalpur, Ahmedabad, Gujarat, India - 380051",
          };
        } else {
          return sendJsonRpc(
            id,
            {
              error: {
                code: -32601,
                message: `Unknown tool: ${name}`,
              },
            },
            200,
          );
        }

        return sendJsonRpc(id, {
          content: [
            {
              type: "text",
              text: JSON.stringify(toolResult, null, 2),
            },
          ],
        });
      }
    }

    // Default fallback response for non-JSONRPC POST
    return res.status(200).json(dynamicManifest);
  }

  // Structured RFC 9457 Error Response
  res.setHeader("Content-Type", "application/problem+json");
  return res.status(405).json({
    type: `${activeDomain}/docs/errors/method-not-allowed`,
    title: "Method Not Allowed",
    status: 405,
    code: "METHOD_NOT_ALLOWED",
    detail: "Only GET and POST methods are supported on the MCP endpoint.",
    instance: "/api/mcp",
    resolution_hint:
      "Send an HTTP GET to fetch the MCP manifest, or an HTTP POST with JSON-RPC 2.0 payload to invoke MCP capabilities.",
  });
}